requirements.txt
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
pandas
tensorboard
fairscale
sentencepiece
gradio==4.19.0
packaging
transformers>=4.43.3
pyyaml
pathlib
peft
Ninja
bitsandbytes
httpx[socks]
einops
regex
h5py
accelerate
pre-commit